+++ src/FolkerKinzel.VCards/md/ReleaseNotes.md
@@ -1,4 +1,4 @@
-- Fixes issues with `RelationProperty` instances that contain `VCard` objects.
+- Fixes an issue that when parsing `X-GENDER` and `X-WAB-GENDER` the group name was not preserved.
 
 &nbsp;
 >**Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.